+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feat: load `.env` files from `$HOME/.gpm` and project directories, if exist, automatically
 - feat: add `--env-file` file to load environment variables from external variables
+- feat: `execute` command which runs shell commands with the environment variables loaded from .env* files
 - fix: exit app if special files could not be loaded
 
 ## 0.10.1